+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/ValidatorApiHandler.java
@@ -101,6 +101,7 @@ import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.validator.BeaconPreparableProposer;
 import tech.pegasys.teku.spec.datastructures.validator.BroadcastValidationLevel;
 import tech.pegasys.teku.spec.datastructures.validator.SubnetSubscription;
+import tech.pegasys.teku.spec.logic.common.util.BlockProposalUtil;
 import tech.pegasys.teku.spec.logic.common.util.SyncCommitteeUtil;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionsGloas;
 import tech.pegasys.teku.statetransition.attestation.AggregatingAttestationPool;
@@ -312,7 +313,8 @@ public class ValidatorApiHandler implements ValidatorApiChannel, SlotEventsChann
   }
 
   @Override
-  public SafeFuture<Optional<ProposerDuties>> getProposerDuties(final UInt64 epoch) {
+  public SafeFuture<Optional<ProposerDuties>> getProposerDuties(
+      final UInt64 epoch, final boolean isFuluCompatible) {
     if (isSyncActive()) {
       return NodeSyncingException.failedFuture();
     }
@@ -328,7 +330,9 @@ public class ValidatorApiHandler implements ValidatorApiChannel, SlotEventsChann
     }
 
     final UInt64 stateSlot =
-        spec.atEpoch(epoch).getBlockProposalUtil().getStateSlotForProposerDuties(spec, epoch);
+        spec.atEpoch(epoch)
+            .getBlockProposalUtil()
+            .getStateSlotForProposerDuties(spec, currentEpoch, epoch);
     LOG.debug(
         "Retrieving proposer duties for epoch {}, current epoch {}, state query slot {}",
         epoch,
@@ -336,7 +340,10 @@ public class ValidatorApiHandler implements ValidatorApiChannel, SlotEventsChann
         stateSlot);
     return combinedChainDataClient
         .getStateAtSlotExact(stateSlot)
-        .thenApply(maybeState -> maybeState.map(state -> getProposerDutiesFromState(state, epoch)));
+        .thenApply(
+            maybeState ->
+                maybeState.map(
+                    state -> getProposerDutiesFromState(state, epoch, isFuluCompatible)));
   }
 
   @Override
@@ -1026,7 +1033,8 @@ public class ValidatorApiHandler implements ValidatorApiChannel, SlotEventsChann
     return !syncStateProvider.getCurrentSyncState().isInSync();
   }
 
-  private ProposerDuties getProposerDutiesFromState(final BeaconState state, final UInt64 epoch) {
+  private ProposerDuties getProposerDutiesFromState(
+      final BeaconState state, final UInt64 epoch, final boolean isFuluCompatible) {
     final List<ProposerDuty> result = getProposalSlotsForEpoch(state, epoch);
     if (spec.atEpoch(epoch).getMilestone().isLessThan(SpecMilestone.FULU)) {
       return new ProposerDuties(
@@ -1036,10 +1044,19 @@ public class ValidatorApiHandler implements ValidatorApiChannel, SlotEventsChann
     }
     final UInt64 currentEpoch = spec.getCurrentEpoch(state);
     final boolean greaterThanCurrentEpoch = epoch.isGreaterThan(currentEpoch);
+    final BlockProposalUtil blockProposalUtil;
+    if (isFuluCompatible) {
+      blockProposalUtil = spec.atEpoch(epoch).getBlockProposalUtil();
+    } else {
+      blockProposalUtil = spec.forMilestone(SpecMilestone.PHASE0).getBlockProposalUtil();
+    }
     final Bytes32 dependentRoot =
-        greaterThanCurrentEpoch
-            ? spec.atEpoch(epoch).getBeaconStateUtil().getCurrentDutyDependentRoot(state)
-            : spec.atEpoch(epoch).getBeaconStateUtil().getPreviousDutyDependentRoot(state);
+        blockProposalUtil.getBlockProposalDependentRoot(
+            combinedChainDataClient.getBestBlockRoot().orElseThrow(),
+            spec.atEpoch(epoch).getBeaconStateUtil().getPreviousDutyDependentRoot(state),
+            spec.atEpoch(epoch).getBeaconStateUtil().getCurrentDutyDependentRoot(state),
+            spec.computeEpochAtSlot(state.getSlot()),
+            epoch);
     LOG.trace(
         "state epoch {}, duties epoch {}, greaterThanCurrentEpoch {}, dependentRoot {}",
         currentEpoch,