+++ ethereum/statetransition/src/test/java/tech/pegasys/teku/statetransition/datacolumns/CustodyGroupCountManagerImplTest.java
@@ -36,7 +36,6 @@ import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.TestSpecFactory;
-import tech.pegasys.teku.spec.config.SpecConfigFulu;
 import tech.pegasys.teku.spec.logic.versions.fulu.helpers.MiscHelpersFulu;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 import tech.pegasys.teku.statetransition.CustodyGroupCountChannel;
@@ -54,7 +53,6 @@ public class CustodyGroupCountManagerImplTest {
   private final StubMetricsSystem metricsSystem = new StubMetricsSystem();
   private MiscHelpersFulu miscHelpersFulu;
   private Spec spec;
-  private SpecConfigFulu specConfigFulu;
   private CustodyGroupCountManagerImpl custodyGroupCountManager;
 
   @Test
@@ -78,7 +76,6 @@ public class CustodyGroupCountManagerImplTest {
     custodyGroupCountManager =
         new CustodyGroupCountManagerImpl(
             spec,
-            spec.getGenesisSpecConfig().toVersionFulu().orElseThrow(),
             spec.getGenesisSpec().miscHelpers().toVersionFulu().orElseThrow(),
             proposersDataManager,
             custodyGroupCountChannel,
@@ -106,7 +103,6 @@ public class CustodyGroupCountManagerImplTest {
     custodyGroupCountManager =
         new CustodyGroupCountManagerImpl(
             spec,
-            spec.getGenesisSpecConfig().toVersionFulu().orElseThrow(),
             spec.getGenesisSpec().miscHelpers().toVersionFulu().orElseThrow(),
             proposersDataManager,
             custodyGroupCountChannel,
@@ -133,7 +129,6 @@ public class CustodyGroupCountManagerImplTest {
     custodyGroupCountManager =
         new CustodyGroupCountManagerImpl(
             spec,
-            spec.getGenesisSpecConfig().toVersionFulu().orElseThrow(),
             spec.getGenesisSpec().miscHelpers().toVersionFulu().orElseThrow(),
             proposersDataManager,
             custodyGroupCountChannel,
@@ -217,7 +212,6 @@ public class CustodyGroupCountManagerImplTest {
                             .balancePerAdditionalCustodyGroup(UInt64.valueOf(32000000000L))
                             .minEpochsForDataColumnSidecarsRequests(64)));
 
-    specConfigFulu = SpecConfigFulu.required(spec.forMilestone(SpecMilestone.FULU).getConfig());
     miscHelpersFulu =
         spy(MiscHelpersFulu.required(spec.forMilestone(SpecMilestone.FULU).miscHelpers()));
 
@@ -226,7 +220,6 @@ public class CustodyGroupCountManagerImplTest {
     custodyGroupCountManager =
         new CustodyGroupCountManagerImpl(
             spec,
-            specConfigFulu,
             miscHelpersFulu,
             proposersDataManager,
             custodyGroupCountChannel,