+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/Eth2P2PNetworkBuilder.java
@@ -159,6 +159,7 @@ public class Eth2P2PNetworkBuilder {
   protected DebugDataDumper debugDataDumper;
   private DasGossipLogger dasGossipLogger;
   private DasReqRespLogger dasReqRespLogger;
+  private Supplier<Boolean> isSuperNodeSupplier;
   private DataColumnSidecarArchiveReconstructor dataColumnSidecarArchiveReconstructor;
 
   protected Eth2P2PNetworkBuilder() {}
@@ -408,7 +409,8 @@ public class Eth2P2PNetworkBuilder {
               debugDataDumper,
               dasGossipLogger,
               executionProofOperationProcessor,
-              config.isExecutionProofTopicEnabled());
+              config.isExecutionProofTopicEnabled(),
+              isSuperNodeSupplier);
       case GLOAS ->
           new GossipForkSubscriptionsGloas(
               forkAndSpecMilestone.getFork(),
@@ -435,7 +437,8 @@ public class Eth2P2PNetworkBuilder {
               debugDataDumper,
               dasGossipLogger,
               executionProofOperationProcessor,
-              config.isExecutionProofTopicEnabled());
+              config.isExecutionProofTopicEnabled(),
+              isSuperNodeSupplier);
     };
   }
 
@@ -469,7 +472,8 @@ public class Eth2P2PNetworkBuilder {
               debugDataDumper,
               dasGossipLogger,
               bpo,
-              config.isExecutionProofTopicEnabled());
+              config.isExecutionProofTopicEnabled(),
+              isSuperNodeSupplier);
       case GLOAS ->
           new GossipForkSubscriptionsGloasBpo(
               forkAndSpecMilestone.getFork(),
@@ -497,7 +501,8 @@ public class Eth2P2PNetworkBuilder {
               debugDataDumper,
               dasGossipLogger,
               bpo,
-              config.isExecutionProofTopicEnabled());
+              config.isExecutionProofTopicEnabled(),
+              isSuperNodeSupplier);
       default ->
           throw new IllegalStateException(
               "BPO is not supported for: " + forkAndSpecMilestone.getSpecMilestone());
@@ -880,6 +885,11 @@ public class Eth2P2PNetworkBuilder {
 
   public Eth2P2PNetworkBuilder reqRespDasLogger(final DasReqRespLogger dasReqRespLogger) {
     this.dasReqRespLogger = dasReqRespLogger;
+    return this;
+  }
+
+  public Eth2P2PNetworkBuilder isSuperNodeSupplier(final Supplier<Boolean> isSuperNodeSupplier) {
+    this.isSuperNodeSupplier = isSuperNodeSupplier;
     return this;
   }