+++ networking/eth2/src/testFixtures/java/tech/pegasys/teku/networking/eth2/Eth2P2PNetworkFactory.java
@@ -185,6 +185,7 @@ public class Eth2P2PNetworkFactory {
     protected Duration eth2StatusUpdateInterval;
     protected Spec spec = TestSpecFactory.createMinimalPhase0();
     protected DebugDataDumper debugDataDumper;
+    protected Supplier<Boolean> isSuperNodeSupplier;
 
     public Eth2P2PNetwork startNetwork() throws Exception {
       setDefaults();
@@ -373,11 +374,7 @@ public class Eth2P2PNetworkFactory {
             .map(
                 forkAndSpecMilestone ->
                     createSubscriptions(
-                        forkAndSpecMilestone,
-                        metricsSystem,
-                        network,
-                        gossipEncoding,
-                        config.isExecutionProofTopicEnabled()))
+                        forkAndSpecMilestone, metricsSystem, network, gossipEncoding, config))
             .forEach(gossipForkManagerBuilder::fork);
 
         final GossipForkManager gossipForkManager = gossipForkManagerBuilder.build();
@@ -406,7 +403,7 @@ public class Eth2P2PNetworkFactory {
         final NoOpMetricsSystem metricsSystem,
         final DiscoveryNetwork<?> network,
         final GossipEncoding gossipEncoding,
-        final boolean isExecutionProofTopicEnabled) {
+        final P2PConfig p2PConfig) {
       return switch (forkAndSpecMilestone.getSpecMilestone()) {
         case PHASE0 ->
             new GossipForkSubscriptionsPhase0(
@@ -520,7 +517,7 @@ public class Eth2P2PNetworkFactory {
                 signedBlsToExecutionChangeProcessor,
                 debugDataDumper,
                 executionProofOperationProcessor,
-                isExecutionProofTopicEnabled);
+                p2PConfig.isExecutionProofTopicEnabled());
         case FULU ->
             new GossipForkSubscriptionsFulu(
                 forkAndSpecMilestone.getFork(),
@@ -544,7 +541,8 @@ public class Eth2P2PNetworkFactory {
                 debugDataDumper,
                 DasGossipLogger.NOOP,
                 executionProofOperationProcessor,
-                isExecutionProofTopicEnabled);
+                p2PConfig.isExecutionProofTopicEnabled(),
+                isSuperNodeSupplier);
         case GLOAS ->
             new GossipForkSubscriptionsGloas(
                 forkAndSpecMilestone.getFork(),
@@ -571,7 +569,8 @@ public class Eth2P2PNetworkFactory {
                 debugDataDumper,
                 DasGossipLogger.NOOP,
                 executionProofOperationProcessor,
-                isExecutionProofTopicEnabled);
+                p2PConfig.isExecutionProofTopicEnabled(),
+                isSuperNodeSupplier);
       };
     }
 
@@ -677,6 +676,9 @@ public class Eth2P2PNetworkFactory {
       }
       if (executionPayloadBidProcessor == null) {
         executionPayloadBidProcessor = OperationProcessor.noop();
+      }
+      if (isSuperNodeSupplier == null) {
+        isSuperNodeSupplier = () -> false;
       }
     }
 
@@ -894,6 +896,12 @@ public class Eth2P2PNetworkFactory {
     public Eth2P2PNetworkBuilder p2pDebugDataDumper(final DebugDataDumper debugDataDumper) {
       checkNotNull(debugDataDumper);
       this.debugDataDumper = debugDataDumper;
+      return this;
+    }
+
+    public Eth2P2PNetworkBuilder isSuperNodeSupplier(final Supplier<Boolean> isSuperNodeSupplier) {
+      checkNotNull(isSuperNodeSupplier);
+      this.isSuperNodeSupplier = isSuperNodeSupplier;
       return this;
     }
   }